+++ src/lib/serialization/index.ts
@@ -5,6 +5,5 @@ export {
     type DeserializerComponent,
 } from "./deserializer";
 export { SerializeEvent } from "./events";
-export type { SerializeEventData } from "./events";
 export * as JSONOutput from "./schema";
 export { Serializer } from "./serializer";